pyproject.toml
[tool.poetry]
name = "haru"
version = "0.1.0"
description = ""
authors = ["Seunghyeon Yang <61305117+1yangsh@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.28"
black = "^24.3.0"
isort = "^5.13.2"
pycln = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

